Detrmine if you have unlocks Seek out a Bird to day by looking at skill description. Also calculate number of times you have used it by examining the mp cost of next cast.
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@19712 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/persistence/DebugDatabase.java
@@ -1992,10 +1992,10 @@ public class DebugDatabase
 	}
 
 	private static final Pattern SKILL_MP_COST_PATTERN = Pattern.compile( "<b>MP Cost:</b> (\\d+)" );
-	public static final int parseSkillMPCost( final String text )
+	public static final long parseSkillMPCost( final String text )
 	{
 		Matcher matcher = DebugDatabase.SKILL_MP_COST_PATTERN.matcher( text );
-		return matcher.find() ? StringUtilities.parseInt( matcher.group( 1 ) ) : 0;
+		return matcher.find() ? StringUtilities.parseLong( matcher.group( 1 ) ) : 0;
 	}
 
 	// Gives Effect: <b><a class=nounder href="desc_effect.php?whicheffect=69dcf3d8fe46c29e7fb6075d06448c95">Your Fifteen Minutes</a>
